+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Fixed
 
+- #1986 fix: trigger screen is usable on slightly rectangular screens with a low DPI
 - Bugs with expert mode auto starting time.
 
 ## [4.0.0 Beta 6](https://github.com/sds100/KeyMapper/releases/tag/v4.0.0-beta.06)